% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Histograms.R
\name{AgeHistogram}
\alias{AgeHistogram}
\title{Histogram of the Ages}
\usage{
AgeHistogram(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The histogram of the age of the faculty
}
\description{
Makes and displays the Histogram of the ages of the faculty
}

