% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\name{AgeRange}
\alias{AgeRange}
\title{Rangee of the ages of the Faculty}
\usage{
AgeRange(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The range age in the faculty
}
\description{
Finds the range of the ages within the faculty
}

