% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\name{AverageAge}
\alias{AverageAge}
\title{Average Age}
\usage{
AverageAge(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The average age of the faculty
}
\description{
Finds the average age of the Faculty
}

