% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanname}
\alias{cleanname}
\title{Cleaning the Data 2.0 - By Name}
\usage{
cleanname(df)
}
\arguments{
\item{data}{frame          The name of the data frame, Faculty (from cleanyr3)}
}
\value{
A new dataframe (unassigned) that now has the name of the Faculty, cleaning out unnecessary symbols that had previously existed in the data.
}
\description{
Cleans the data for the Names of the faculty.
}

