% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleansubject.R
\name{cleansubject}
\alias{cleansubject}
\title{Cleaning the Data 3.0 - By Department}
\usage{
cleansubject(df)
}
\arguments{
\item{data}{frame          The name of the data frame, Faculty (from cleanname)}
}
\value{
A new dataframe (unassigned) that now has department of the Faculty member, cleaning out unnecessary symbols and terminology that had previously existed in the data.
}
\description{
Cleans the data for the Department the faculty is a member of.
}

