% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanyr1}
\alias{cleanyr1}
\title{Cleaning the Data 1.1 - By Year}
\usage{
cleanyr1(df)
}
\arguments{
\item{data}{frame          The name of the data frame, Faculty (from upload)}
}
\value{
A new dataframe (unassigned) that now has all years at which people received their degrees, in the 3rd column. Some frames may include NA's.
}
\description{
Cleans the data for the Year people received their degrees by recursively deleting frames in the 3rd column that is not a year, eventually shifting all years not in the 3rd column to the 3rd column, while introducing NA's (which will be later removed) as to not disrupt the index of the dataframe.
}

