% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanyr2}
\alias{cleanyr2}
\title{Cleaning the Data 1.2 - By Year}
\usage{
cleanyr2(df)
}
\arguments{
\item{data}{frame          The name of the data frame, Faculty (from cleanyr1)}
}
\value{
A new dataframe (unassigned) that now has all years at which people received their degrees, in the 3rd column. Some frames may still be blank.
}
\description{
Cleans the data for the Year people received their degrees by recursively deleting frames in the 3rd column that contains NA's due to the introduction of NA's when removing unwanted data in cleanyr1()
}

