% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanyr.R
\name{cleanyr3}
\alias{cleanyr3}
\title{Cleaning the Data 1.3 - By Year}
\usage{
cleanyr3(df)
}
\arguments{
\item{data}{frame          The name of the data frame, Faculty (from cleanyr2)}
}
\value{
A new dataframe (unassigned) that now has all years at which people received their degrees, in the 3rd column.
}
\description{
Cleans the data for the Year people received their degrees by recursively deleting frames in the 3rd column that is blank.
}

