% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\name{MaxAge}
\alias{MaxAge}
\title{Maximum Age of the Faculty}
\usage{
MaxAge(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The Maximum age in the faculty
}
\description{
Finds the Maximum age within the faculty
}

