% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\name{MedianAge}
\alias{MedianAge}
\title{Median Age of the Faculty}
\usage{
MedianAge(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The Median age in the faculty
}
\description{
Finds the Median age within the faculty
}

