% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\name{MinAge}
\alias{MinAge}
\title{Minimum Age of the Faculty}
\usage{
MinAge(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The Minimum age in the faculty
}
\description{
Finds the Minimum age within the faculty
}

