% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Summary of the Faculty}
\usage{
Summary(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame, Faculty}
}
\value{
The the summary data of the ages of the faculty
}
\description{
Displays the minimum, 1st quartile, median, 3rd quartile, and maximum of the ages of the faculty.
}

