% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upload.R
\name{upload}
\alias{upload}
\title{Uploading the Data}
\usage{
upload(txtfile)
}
\arguments{
\item{.txt}{file          The name of the txt file}
}
\value{
The variable "Faculty", containing the data table that is imported
}
\description{
Reads in a txt file in table format and creates a data frame from it, with cases corresponding to lines and variables to fileds in the file.
}

