% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Histograms.R
\name{YearHistogram}
\alias{YearHistogram}
\title{Histogram of the Years the faculty received their degrees}
\usage{
YearHistogram(df)
}
\arguments{
\item{data}{frame          The name of the clean data frame}
}
\value{
The histogram of the year the faculty received their degrees
}
\description{
Makes and displays the Histogram of the years the faculty received their degrees
}

